feat[#17] : withTooltip hoc 제거하고 컴포넌트로 변경
