+++ app/src/features/snippet/ui/DownLoadSnippetToSVGButton.tsx
@@ -1,8 +1,6 @@
-import withTooltip from "@/shared/lib/withTooltip";
 import { toSvg } from "html-to-image";
-import { SNIPPET_TOOLTIP } from "../config";
 
-export const DownLoadSnippetToSVGButton = withTooltip(() => {
+export const DownLoadSnippetToSVGButton = () => {
   const handleDownload = async () => {
     const $codeBlock = document.querySelector("#codeBlock") as HTMLDivElement;
 
@@ -40,4 +38,4 @@ export const DownLoadSnippetToSVGButton = withTooltip(() => {
       Download Image
     </button>
   );
-})(SNIPPET_TOOLTIP.downloadSnippet, 1);
+};
